+++ global.d.ts
@@ -1,0 +1,3 @@
+declare module 'vecn' {
+  export function vec3(any, any, any)
+}