+++ src/components/ViewportScrollScene.js
@@ -7,7 +7,6 @@ import { useCanvasStore } from '../store'
 import useScrollRig from '../hooks/useScrollRig'
 import DebugMesh from './DebugMesh'
 import useTracker from '../hooks/useTracker'
-import useScrollbar from '../scrollbar/useScrollbar'
 
 /**
  * Generic THREE.js Scene that tracks the dimensions and position of a DOM element while scrolling
@@ -37,15 +36,15 @@ let ViewportScrollScene = ({
   // const setEvents = useThree((state) => state.setEvents)
 
   const { renderViewport } = useScrollRig()
-  const { scroll } = useScrollbar()
 
   const pageReflow = useCanvasStore((state) => state.pageReflow)
   const scaleMultiplier = useCanvasStore((state) => state.scaleMultiplier)
 
-  const { update, bounds, scale, position, scrollState, inViewport } = useTracker(
-    { track, rootMargin: inViewportMargin, threshold: inViewportThreshold },
-    [pageReflow, scene]
-  )
+  const { rect, bounds, scale, position, scrollState, inViewport } = useTracker({
+    track,
+    rootMargin: inViewportMargin,
+    threshold: inViewportThreshold,
+  })
 
   // Hide scene when outside of viewport if `hideOffscreen` or set to `visible` prop
   useLayoutEffect(() => {
@@ -56,12 +55,12 @@ let ViewportScrollScene = ({
 
   // Find bounding box & scale mesh on resize
   useLayoutEffect(() => {
-    const viewportWidth = bounds.width * scaleMultiplier
-    const viewportHeight = bounds.height * scaleMultiplier
+    const viewportWidth = rect.width * scaleMultiplier
+    const viewportHeight = rect.height * scaleMultiplier
     const cameraDistance = Math.max(viewportWidth, viewportHeight)
     setCameraDistance(cameraDistance)
 
-    // Calculate FOV to match the DOM bounds for this camera distance
+    // Calculate FOV to match the DOM rect for this camera distance
     if (camera.current && !orthographic) {
       camera.current.aspect =
         (viewportWidth + margin * 2 * scaleMultiplier) / (viewportHeight + margin * 2 * scaleMultiplier)
@@ -74,7 +73,7 @@ let ViewportScrollScene = ({
     }
     // trigger a frame
     invalidate()
-  }, [track, pageReflow, bounds, scaleMultiplier])
+  }, [track, pageReflow, rect, scaleMultiplier])
 
   const compute = React.useCallback(
     (event, state) => {
@@ -83,8 +82,8 @@ let ViewportScrollScene = ({
         const { width, height, left, top } = bounds
         const mWidth = width + margin * 2
         const mHeight = height + margin * 2
-        const x = event.clientX - left + margin + scroll.x
-        const y = event.clientY - top + margin + scroll.y
+        const x = event.clientX - left + margin
+        const y = event.clientY - top + margin
         state.pointer.set((x / mWidth) * 2 - 1, -(y / mHeight) * 2 + 1)
         state.raycaster.setFromCamera(state.pointer, camera.current)
       }
@@ -106,9 +105,6 @@ let ViewportScrollScene = ({
   useFrame(({ gl }) => {
     if (!scene || !scale) return
 
-    // update element tracker
-    update()
-
     // Render scene to viewport using local camera and limit updates using scissor test
     // Performance improvement - faster than always rendering full canvas
     if (scene.visible) {
@@ -117,7 +113,7 @@ let ViewportScrollScene = ({
         scene,
         camera: camera.current,
         left: bounds.left - margin,
-        top: position.positiveYUpBottom - margin,
+        top: bounds.positiveYUpBottom - margin,
         width: bounds.width + margin * 2,
         height: bounds.height + margin * 2,
       })
@@ -174,7 +170,7 @@ let ViewportScrollScene = ({
         </group>
       </>,
       scene,
-      { events: { compute, priority }, size: { width: bounds.width, height: bounds.height } }
+      { events: { compute, priority }, size: { width: rect.width, height: rect.height } }
     )
   )
 }